#285: invert PeerTransferComponent state into PeerTransferRepository
