+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/di/AppContainer.kt
@@ -10,7 +10,14 @@ import com.tubetoast.tether.network.TransferActivityTracker
 import com.tubetoast.tether.preferences.FileTransferPreferences
 import com.tubetoast.tether.preferences.PeerPreferencesStore
 import com.tubetoast.tether.presentation.RootComponentFactory
+import com.tubetoast.tether.presentation.transfer.PeerTransferRepository
+import com.tubetoast.tether.presentation.transfer.PeerTransferRepositoryImpl
 import com.tubetoast.tether.security.TrustedDeviceStore
+import com.tubetoast.tether.transfer.ReceiveEvent
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.flow.MutableSharedFlow
 
 abstract class AppContainer {
     protected abstract val namePersistence: DeviceNamePersistence
@@ -21,7 +28,20 @@ abstract class AppContainer {
     open val transferActivityTracker: TransferActivityTracker = DefaultTransferActivityTracker()
     open val fileClient: FileClient by lazy { FileClient.default(transferActivityTracker) }
     abstract val trustedDeviceStore: TrustedDeviceStore
-    open val rootComponentFactory: RootComponentFactory by lazy { RootComponentFactory(mdnsDiscovery) }
+    open val rootComponentFactory: RootComponentFactory by lazy {
+        RootComponentFactory(mdnsDiscovery, peerTransferRepository)
+    }
     abstract val peerPreferencesStore: PeerPreferencesStore
     abstract val fileTransferPreferences: FileTransferPreferences
+    open val appScope: CoroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Default)
+
+    open val inboundEvents: MutableSharedFlow<ReceiveEvent> = MutableSharedFlow(extraBufferCapacity = 64)
+
+    open val peerTransferRepository: PeerTransferRepository by lazy {
+        PeerTransferRepositoryImpl(
+            batchSenderFactory = { error("TODO(#191): wire BatchSender to FileClient in Phase B") },
+            inboundEvents = inboundEvents,
+            scope = appScope,
+        )
+    }
 }
