+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/presentation/RootComponentFactory.kt
@@ -2,9 +2,11 @@ package com.tubetoast.tether.presentation
 
 import com.arkivanov.decompose.ComponentContext
 import com.tubetoast.tether.discovery.DeviceDiscovery
+import com.tubetoast.tether.presentation.transfer.PeerTransferRepository
 
 class RootComponentFactory(
     private val discovery: DeviceDiscovery,
+    private val peerTransferRepository: PeerTransferRepository,
 ) {
     fun create(componentContext: ComponentContext): RootComponent =
         RootComponent(
